chore(tooling): full react-doctor triage, score 61 -> 75
Full-scope npx react-doctor@latest scan had drifted to score 61 (599
errors before test-file/generated-fixture ignores were even applied).
Fixes every finding that isn't auth code or a genuine design-judgment
call:

- deslop/unused-export and deslop/unused-file turned off project-wide
  in doctor.config.json - both rules can't resolve the @app/* tsconfig
  alias or platform-variant siblings (.ios/.web/.android), producing
  40-60+ false positives on every full scan across several prior
  triage passes. Documented in AGENTS.md. Found and removed the 4
  exports that were genuinely dead this time.
- anti-slop findings in scripts/workflows/*.ts, github-actions.ts, and
  player-website/src/scripts/player.js - real fixes (parsed CLI/env
  input instead of typeof narrowing, satisfies over widened Record
  types, safety comments where an assertion is unavoidable), not
  suppressions; oxlint's app-focused sweep never scans these paths but
  react-doctor's full scan does.
- React Compiler try/finally bails fixed in 9 more hooks/services
  (drop finally, move cleanup after try/catch unconditionally, or
  .finally() on a promise chain), plus two new unsupported-syntax
  bails (getter-style ref, ??= compound assignment).
- 3 of 5 no-giant-component warnings resolved via mechanical,
  behavior-preserving component extraction (ChatPreferenceDefaultContent,
  ChatPreferenceForm, SettingsIndexScreen); StreamPlayer.tsx and
  LiveStreamScreen.tsx left deferred - their length comes from
  genuinely interdependent state that needs a design decision, not a
  scripted split.
- A handful of new scoped, justified doctor.config.json overrides
  (useLazyRef.ts lazy-init pattern, twitch-chat-service.ts purity,
  EmoteActionSheet.tsx refs, ImageBenchmarkScreen.tsx) - each checked
  against the code and documented in AGENTS.md, none a blanket rule-off.
- Various one-off correctness fixes: a stuck-loading-flag bug
  (usePinnedChatMessage), an O(n) array scan in a loop, a
  spread-after-defaults clobber, a stale-prop-sync effect, unstable
  list keys, and an unpinned installer script in a GitHub workflow.

Auth code and its tests remain untouched per standing policy.

bun run ci:local is green end to end.
