+++ src/components/StreamPlayer/StreamPlayer.tsx
@@ -18,90 +18,18 @@ import { DebugErrorOverlay, TouchBlockOverlay } from './StreamPlayerOverlays';
 import { StreamPlayerPoster } from './StreamPlayerPoster';
 import { StreamPlayerWebView } from './StreamPlayerWebView';
 import { buildRawTwitchPlayerUrl } from './twitchPlayerSource/buildRawTwitchPlayerUrl';
-import { buildTwitchAutoplayEnsureScript } from './twitchPlayerSource/buildTwitchAutoplayEnsureScript';
-import { buildTwitchCaptionHiderScript } from './twitchPlayerSource/buildTwitchCaptionHiderScript';
-import { buildTwitchChromeHiderScript } from './twitchPlayerSource/buildTwitchChromeHiderScript';
+import {
+  buildStreamPlayerInjectedJavaScript,
+  buildTwitchAuthHelperScript,
+} from './twitchPlayerSource/buildStreamPlayerInjectedJavaScript';
 import { buildTwitchClipPlayerUrl } from './twitchPlayerSource/buildTwitchClipPlayerUrl';
-import { buildTwitchContentGateAcceptScript } from './twitchPlayerSource/buildTwitchContentGateAcceptScript';
-import { buildTwitchContentGateWatcherScript } from './twitchPlayerSource/buildTwitchContentGateWatcherScript';
-import { buildTwitchEmbedErrorWatcherScript } from './twitchPlayerSource/buildTwitchEmbedErrorWatcherScript';
-import { buildTwitchLatencyTrackerScript } from './twitchPlayerSource/buildTwitchLatencyTrackerScript';
-import { buildTwitchLiveSyncScript } from './twitchPlayerSource/buildTwitchLiveSyncScript';
-import { buildTwitchPipBridgeScript } from './twitchPlayerSource/buildTwitchPipBridgeScript';
 import { buildTwitchPlayerAudioDefaultScript } from './twitchPlayerSource/buildTwitchPlayerAudioDefaultScript';
 import { buildTwitchPlayerQualityDefaultScript } from './twitchPlayerSource/buildTwitchPlayerQualityDefaultScript';
-import { buildTwitchPlayerStateScript } from './twitchPlayerSource/buildTwitchPlayerStateScript';
 import type { StreamPlayerProps } from './types';
 import { usePlayerBridge } from './usePlayerBridge';
 import { useStreamPlayerControls } from './useStreamPlayerControls';
 
 export type { StreamInfo, StreamPlayerProps, StreamPlayerRef } from './types';
-
-const TWITCH_AUTH_HELPER_SCRIPT = `
-(() => {
-  const post = type => {
-    try {
-      window.ReactNativeWebView?.postMessage(JSON.stringify({ type }));
-    } catch {}
-  };
-
-  window.open = url => {
-    if (typeof url === 'string' && url.length > 0) {
-      window.location.assign(url);
-    }
-    return window;
-  };
-
-  let postedAuthComplete = false;
-  const detectAuthComplete = () => {
-    if (postedAuthComplete || !document.body) {
-      return;
-    }
-
-    const text = document.body.textContent?.toLowerCase() ?? '';
-    if (
-      (text.includes("you're logged in") || text.includes("you’re logged in")) &&
-      text.includes('refresh the page')
-    ) {
-      postedAuthComplete = true;
-      post('twitchAuthComplete');
-    }
-  };
-
-  detectAuthComplete();
-  new MutationObserver(detectAuthComplete).observe(document.documentElement, {
-    childList: true,
-    subtree: true});
-})();
-true;
-`;
-
-/**
- * Polls the VOD <video> element's position and reports it to native so the
- * last-known offset survives a WebView reload. The stock player owns the
- * scrubber; we only observe, so this never fights the user's own seeks.
- */
-const VOD_PROGRESS_TRACKER_SCRIPT = `
-(() => {
-  if (window.__foamVodProgressInstalled) {
-    return;
-  }
-  window.__foamVodProgressInstalled = true;
-
-  setInterval(() => {
-    try {
-      const video = document.querySelector('video');
-      const time = video ? video.currentTime : 0;
-      if (Number.isFinite(time) && time > 0) {
-        window.ReactNativeWebView?.postMessage(
-          JSON.stringify({ type: 'vodProgress', payload: { currentTime: time } }),
-        );
-      }
-    } catch {}
-  }, 3000);
-})();
-true;
-`;
 
 /**
  * The unscripted live player gives no 'playing' bridge event, so the poster is
@@ -247,7 +175,9 @@ export const StreamPlayer = memo(function StreamPlayer({
   const sourceKey = `${channel ?? ''}|${clip ?? ''}|${video ?? ''}|${embedParent}|${autoplay}|${initialMuted}|${deferOverlayUntilUserUnmute}`;
 
   const generation = `${sourceKey}|${webViewKey}`;
-  generationRef.current = generation;
+  useEffect(() => {
+    generationRef.current = generation;
+  }, [generation]);
   const isPlayerLoading = loadedGeneration !== generation;
 
   /**
@@ -400,50 +330,13 @@ export const StreamPlayer = memo(function StreamPlayer({
     [clip, channelName, video, embedParent, autoplay, initialMuted, webViewKey],
   );
 
-  /**
-   * The stock player.twitch.tv page runs unscripted: Twitch's own UI handles
-   * playback, so the playerControls bootstrap is not injected and the bridge
-   * (ready/playing/pause messages, native controls overlay) stays dormant.
-   * We auto-accept the mature-content gate, hide the text track ('hidden', not
-   * 'disabled', which would stall WKWebView's native HLS AVPlayer), and — when
-   * autoplaying — nudge the video into unmuted playback since Twitch's embed
-   * otherwise tends to land paused/muted. The player's own chrome is hidden
-   * only when foam is drawing its own overlay controls over the video.
-   */
-  const injectedJavaScript =
-    TWITCH_AUTH_HELPER_SCRIPT +
-    '\n' +
-    // Detect Twitch's "this embed is misconfigured" error page (bad parent) on
-    // every player kind so a broken embed reports to Sentry instead of only
-    // timing out.
-    buildTwitchEmbedErrorWatcherScript() +
-    '\n' +
-    buildTwitchContentGateAcceptScript() +
-    '\n' +
-    buildTwitchCaptionHiderScript() +
-    (autoplay && !clip
-      ? '\n' + buildTwitchAutoplayEnsureScript({ muted: initialMuted })
-      : '') +
-    (showOverlayControls && !clip
-      ? '\n' +
-        buildTwitchChromeHiderScript() +
-        '\n' +
-        buildTwitchPlayerStateScript() +
-        '\n' +
-        buildTwitchContentGateWatcherScript()
-      : '') +
-    // Live + custom-player only: read broadcaster latency for the chat pill and seek to live at start.
-    (showOverlayControls && !clip && !video
-      ? '\n' +
-        buildTwitchLatencyTrackerScript() +
-        '\n' +
-        buildTwitchLiveSyncScript({})
-      : '') +
-    (video ? '\n' + VOD_PROGRESS_TRACKER_SCRIPT : '') +
-    // iOS-only: WKWebView is the only WebView with a presentation-mode API.
-    (PIP_ENABLED && Platform.OS === 'ios' && !clip
-      ? '\n' + buildTwitchPipBridgeScript()
-      : '');
+  const injectedJavaScript = buildStreamPlayerInjectedJavaScript({
+    autoplay,
+    clip,
+    initialMuted,
+    showOverlayControls,
+    video,
+  });
 
   /**
    * The tracker posts unsolicited `vodProgress` messages; capture those for
@@ -473,7 +366,7 @@ export const StreamPlayer = memo(function StreamPlayer({
   );
 
   const injectedJavaScriptBeforeContentLoaded = clip
-    ? TWITCH_AUTH_HELPER_SCRIPT
+    ? buildTwitchAuthHelperScript()
     : buildTwitchPlayerQualityDefaultScript({
         defaultQuality: '720p60',
         maxBitrateBps: 3_500_000,
