+++ src/components/StreamPlayer/StreamPlayerPoster.tsx
@@ -33,9 +33,14 @@ export const StreamPlayerPoster = memo(function StreamPlayerPoster({
   const [rendered, setRendered] = useState(visible);
   const opacity = useSharedValue(visible ? 1 : 0);
 
+  // Adjust synchronously during render (rather than in the effect below) so a
+  // remount-to-visible never paints one frame of the stale hidden state first.
+  if (visible && !rendered) {
+    setRendered(true);
+  }
+
   useEffect(() => {
     if (visible) {
-      setRendered(true);
       opacity.set(1);
       return;
     }
