+++ src/screens/Preferences/ChatPreferenceForm.tsx
@@ -1,60 +1,22 @@
-import { type ReactElement, useMemo } from 'react';
-import { StyleSheet, useWindowDimensions, View } from 'react-native';
+import { useMemo } from 'react';
+import { StyleSheet, useWindowDimensions } from 'react-native';
 
-import {
-  Button,
-  Form,
-  Host,
-  Picker,
-  RNHostView,
-  Section,
-  Text as NativeText,
-  Toggle,
-} from '@expo/ui/swift-ui';
-import { tag } from '@expo/ui/swift-ui/modifiers';
-import { router } from 'expo-router';
+import { Form, Host } from '@expo/ui/swift-ui';
 
 import { usePreferences } from '@app/store/preferenceStore';
 import { theme } from '@app/styles/themes';
-import {
-  EMOJI_STYLE_OPTIONS,
-  getEmojiEmotes,
-} from '@app/utils/emoji/emojiEmotes';
+import { getEmojiEmotes } from '@app/utils/emoji/emojiEmotes';
 
-import {
-  DensityPreview,
-  EmojiStylePreview,
-  PreviewLabel,
-  ProviderPreviewItem,
-} from './ChatPreferencePreviewWidgets';
-import { ChatPreferencePreview } from './ChatPreferencesPreview';
-import {
-  CHAT_DELAY_OPTIONS,
-  DELETED_STYLE_OPTIONS,
-  DENSITY_OPTIONS,
-  EMOJI_PREVIEW_SHORTCODES,
-  FONT_SCALE_OPTIONS,
-  type PreviewProvider,
-  SCROLLBACK_OPTIONS,
-  TIMESTAMP_FORMAT_OPTIONS,
-} from './chatPreferenceTypes';
-
-type ProviderSection = {
-  title: string;
-  provider: PreviewProvider;
-  emotes: boolean;
-  badges: boolean;
-  onEmotes: (value: boolean) => void;
-  onBadges: (value: boolean) => void;
-};
-
-function hostPreview(node: ReactElement, width: number, padded = true) {
-  return (
-    <RNHostView matchContents>
-      <View style={[{ width }, padded ? styles.previewRow : null]}>{node}</View>
-    </RNHostView>
-  );
-}
+import { ChatPreferenceFormContextSection } from './ChatPreferenceFormContextSection';
+import { ChatPreferenceFormEmojiSection } from './ChatPreferenceFormEmojiSection';
+import { ChatPreferenceFormHighlightsSection } from './ChatPreferenceFormHighlightsSection';
+import { ChatPreferenceFormLayoutSection } from './ChatPreferenceFormLayoutSection';
+import { ChatPreferenceFormMediaSection } from './ChatPreferenceFormMediaSection';
+import { ChatPreferenceFormModerationSection } from './ChatPreferenceFormModerationSection';
+import { ChatPreferenceFormPerformanceSection } from './ChatPreferenceFormPerformanceSection';
+import { ChatPreferenceFormProviderSections } from './ChatPreferenceFormProviderSections';
+import { ChatPreferenceFormSyncSection } from './ChatPreferenceFormSyncSection';
+import { EMOJI_PREVIEW_SHORTCODES } from './chatPreferenceTypes';
 
 export function ChatPreferenceForm() {
   const preferences = usePreferences();
@@ -78,335 +40,60 @@ export function ChatPreferenceForm() {
     showUnreadJumpPill: preferences.showUnreadJumpPill,
   };
 
-  const providerSections: ProviderSection[] = [
-    {
-      title: '7TV',
-      provider: '7tv',
-      emotes: preferences.show7TvEmotes,
-      badges: preferences.show7tvBadges,
-      onEmotes: (value: boolean) => update({ show7TvEmotes: value }),
-      onBadges: (value: boolean) => update({ show7tvBadges: value }),
-    },
-    {
-      title: 'BTTV',
-      provider: 'bttv',
-      emotes: preferences.showBttvEmotes,
-      badges: preferences.showBttvBadges,
-      onEmotes: (value: boolean) => update({ showBttvEmotes: value }),
-      onBadges: (value: boolean) => update({ showBttvBadges: value }),
-    },
-    {
-      title: 'FFZ',
-      provider: 'ffz',
-      emotes: preferences.showFFzEmotes,
-      badges: preferences.showFFzBadges,
-      onEmotes: (value: boolean) => update({ showFFzEmotes: value }),
-      onBadges: (value: boolean) => update({ showFFzBadges: value }),
-    },
-    {
-      title: 'Twitch',
-      provider: 'twitch',
-      emotes: preferences.showTwitchEmotes,
-      badges: preferences.showTwitchBadges,
-      onEmotes: (value: boolean) => update({ showTwitchEmotes: value }),
-      onBadges: (value: boolean) => update({ showTwitchBadges: value }),
-    },
-  ];
-
   return (
     <Host style={styles.host}>
       <Form>
-        <Section title='Layout'>
-          <Picker
-            label='Message Density'
-            systemImage='list.bullet'
-            selection={preferences.chatDensity}
-            onSelectionChange={value => update({ chatDensity: value })}
-          >
-            {DENSITY_OPTIONS.map(option => (
-              <NativeText key={option.value} modifiers={[tag(option.value)]}>
-                {option.label}
-              </NativeText>
-            ))}
-          </Picker>
-          {hostPreview(
-            <DensityPreview density={preferences.chatDensity} />,
-            previewWidth,
-          )}
-          <Picker
-            label='Font Size'
-            systemImage='textformat.size'
-            selection={preferences.chatFontScale}
-            onSelectionChange={value => update({ chatFontScale: value })}
-          >
-            {FONT_SCALE_OPTIONS.map(option => (
-              <NativeText key={option.value} modifiers={[tag(option.value)]}>
-                {option.label}
-              </NativeText>
-            ))}
-          </Picker>
-          {hostPreview(
-            <ChatPreferencePreview
-              variant='fontScale'
-              value={preferences.chatFontScale}
-            />,
-            previewWidth,
-          )}
-          <Toggle
-            label='Alternating Rows'
-            systemImage='line.3.horizontal'
-            isOn={preferences.showAlternatingChatRows}
-            onIsOnChange={value => update({ showAlternatingChatRows: value })}
-          />
-          {hostPreview(
-            <ChatPreferencePreview
-              variant='alternatingRows'
-              value={preferences.showAlternatingChatRows}
-            />,
-            previewWidth,
-          )}
-          <Toggle
-            label='New Message Animation'
-            systemImage='arrow.up.message'
-            isOn={preferences.animate}
-            onIsOnChange={value => update({ animate: value })}
-          />
-        </Section>
+        <ChatPreferenceFormLayoutSection
+          preferences={preferences}
+          previewWidth={previewWidth}
+          update={update}
+        />
 
-        <Section title='Emoji Style'>
-          <Picker
-            label='Emoji Set'
-            systemImage='face.smiling'
-            selection={preferences.emojiStyle}
-            onSelectionChange={value => update({ emojiStyle: value })}
-          >
-            {EMOJI_STYLE_OPTIONS.map(option => (
-              <NativeText key={option.value} modifiers={[tag(option.value)]}>
-                {option.label}
-              </NativeText>
-            ))}
-          </Picker>
-          {hostPreview(
-            <EmojiStylePreview emotes={emojiPreviewEmotes} />,
-            previewWidth,
-          )}
-        </Section>
+        <ChatPreferenceFormEmojiSection
+          emojiPreviewEmotes={emojiPreviewEmotes}
+          preferences={preferences}
+          previewWidth={previewWidth}
+          update={update}
+        />
 
-        <Section title='Context'>
-          <Toggle
-            label='Historical Recent Messages'
-            systemImage='clock.arrow.circlepath'
-            isOn={preferences.showRecentMessages !== false}
-            onIsOnChange={value => update({ showRecentMessages: value })}
-          />
-          <Toggle
-            label='Show Timestamps'
-            systemImage='clock'
-            isOn={preferences.chatTimestamps}
-            onIsOnChange={value => update({ chatTimestamps: value })}
-          />
-          <Toggle
-            label='Highlight Own Mentions'
-            systemImage='at'
-            isOn={preferences.highlightOwnMentions}
-            onIsOnChange={value => update({ highlightOwnMentions: value })}
-          />
-          <Toggle
-            label='Inline Reply Context'
-            systemImage='arrowshape.turn.up.left'
-            isOn={preferences.showInlineReplyContext}
-            onIsOnChange={value => update({ showInlineReplyContext: value })}
-          />
-          <Toggle
-            label='Show Jump Pill'
-            systemImage='arrow.down.circle'
-            isOn={preferences.showUnreadJumpPill}
-            onIsOnChange={value => update({ showUnreadJumpPill: value })}
-          />
-          <Picker
-            label='Timestamp Format'
-            systemImage='clock.badge'
-            selection={preferences.chatTimestampFormat}
-            onSelectionChange={value => update({ chatTimestampFormat: value })}
-          >
-            {TIMESTAMP_FORMAT_OPTIONS.map(option => (
-              <NativeText key={option.value} modifiers={[tag(option.value)]}>
-                {option.label}
-              </NativeText>
-            ))}
-          </Picker>
-          {hostPreview(
-            <View>
-              <PreviewLabel />
-              <View style={styles.previewSpacer}>
-                <ChatPreferencePreview
-                  variant='context'
-                  value={contextPreview}
-                />
-              </View>
-            </View>,
-            previewWidth,
-          )}
-        </Section>
+        <ChatPreferenceFormContextSection
+          contextPreview={contextPreview}
+          preferences={preferences}
+          previewWidth={previewWidth}
+          update={update}
+        />
 
-        <Section
-          title='Sync'
-          footer={
-            <NativeText>
-              Delay chat so it lines up with the video. Auto matches the
-              measured stream latency.
-            </NativeText>
-          }
-        >
-          <Picker
-            label='Chat Delay'
-            systemImage='timer'
-            selection={String(preferences.chatDelay)}
-            onSelectionChange={value => {
-              const option = CHAT_DELAY_OPTIONS.find(
-                item => String(item.value) === value,
-              );
-              if (option) {
-                update({ chatDelay: option.value });
-              }
-            }}
-          >
-            {CHAT_DELAY_OPTIONS.map(option => (
-              // SwiftUI tag matching needs one type; the values mix 'auto'/'off' and numbers.
-              <NativeText
-                key={option.value}
-                modifiers={[tag(String(option.value))]}
-              >
-                {option.label}
-              </NativeText>
-            ))}
-          </Picker>
-        </Section>
+        <ChatPreferenceFormSyncSection
+          preferences={preferences}
+          update={update}
+        />
 
-        <Section
-          title='Highlights'
-          footer={
-            <NativeText>
-              Highlighted phrases tint matching messages. Mention feedback also
-              buzzes when a highlight matches.
-            </NativeText>
-          }
-        >
-          <Button
-            label='Highlighted Phrases'
-            systemImage='highlighter'
-            onPress={() => router.push('/tabs/settings/chat-highlights')}
-          />
-          <Toggle
-            label='Mention Feedback'
-            systemImage='hand.tap'
-            isOn={preferences.chatMentionHaptics !== false}
-            onIsOnChange={value => update({ chatMentionHaptics: value })}
-          />
-        </Section>
+        <ChatPreferenceFormHighlightsSection
+          preferences={preferences}
+          update={update}
+        />
 
-        <Section title='Moderation'>
-          <Picker
-            label='Deleted Messages'
-            systemImage='trash.slash'
-            selection={preferences.deletedMessageStyle}
-            onSelectionChange={value => update({ deletedMessageStyle: value })}
-          >
-            {DELETED_STYLE_OPTIONS.map(option => (
-              <NativeText key={option.value} modifiers={[tag(option.value)]}>
-                {option.label}
-              </NativeText>
-            ))}
-          </Picker>
-          <Toggle
-            label='Keep History on Clear'
-            systemImage='clock.arrow.circlepath'
-            isOn={preferences.ignoreClearChat === true}
-            onIsOnChange={value => update({ ignoreClearChat: value })}
-          />
-        </Section>
+        <ChatPreferenceFormModerationSection
+          preferences={preferences}
+          update={update}
+        />
 
-        <Section
-          title='Performance'
-          footer={
-            <NativeText>
-              Longer scrollback keeps more messages in memory; 200 is easier on
-              older devices.
-            </NativeText>
-          }
-        >
-          <Picker
-            label='Scrollback'
-            systemImage='text.line.last.and.arrowtriangle.forward'
-            selection={preferences.chatScrollback}
-            onSelectionChange={value => update({ chatScrollback: value })}
-          >
-            {SCROLLBACK_OPTIONS.map(option => (
-              <NativeText key={option.value} modifiers={[tag(option.value)]}>
-                {option.label}
-              </NativeText>
-            ))}
-          </Picker>
-        </Section>
+        <ChatPreferenceFormPerformanceSection
+          preferences={preferences}
+          update={update}
+        />
 
-        {providerSections.map(section => (
-          <Section key={section.title} title={section.title}>
-            <Toggle
-              label='Emotes'
-              systemImage='face.smiling'
-              isOn={section.emotes}
-              onIsOnChange={section.onEmotes}
-            />
-            {hostPreview(
-              <ProviderPreviewItem
-                enabled={section.emotes}
-                provider={section.provider}
-                variant='emotes'
-              />,
-              previewWidth,
-              false,
-            )}
-            <Toggle
-              label='Badges'
-              systemImage='rosette'
-              isOn={section.badges}
-              onIsOnChange={section.onBadges}
-            />
-            {hostPreview(
-              <ProviderPreviewItem
-                enabled={section.badges}
-                provider={section.provider}
-                variant='badges'
-              />,
-              previewWidth,
-              false,
-            )}
-          </Section>
-        ))}
+        <ChatPreferenceFormProviderSections
+          preferences={preferences}
+          previewWidth={previewWidth}
+          update={update}
+        />
 
-        <Section
-          title='Media'
-          footer={
-            <NativeText>
-              Animated Twitch, BTTV, FFZ, and 7TV emotes will render as still
-              images when this is enabled.
-            </NativeText>
-          }
-        >
-          <Toggle
-            label='Disable Emote Animations'
-            systemImage='slash.circle'
-            isOn={preferences.disableEmoteAnimations}
-            onIsOnChange={value => update({ disableEmoteAnimations: value })}
-          />
-          {hostPreview(
-            <ChatPreferencePreview
-              variant='emoteAnimations'
-              value={preferences.disableEmoteAnimations}
-            />,
-            previewWidth,
-          )}
-        </Section>
+        <ChatPreferenceFormMediaSection
+          preferences={preferences}
+          previewWidth={previewWidth}
+          update={update}
+        />
       </Form>
     </Host>
   );
@@ -415,12 +102,5 @@ export function ChatPreferenceForm() {
 const styles = StyleSheet.create({
   host: {
     flex: 1,
-  },
-  previewRow: {
-    paddingHorizontal: theme.space16,
-    paddingVertical: theme.space8,
-  },
-  previewSpacer: {
-    marginTop: theme.space8,
   },
 });
